Fix jl2005a camlib compilation warnings.
* jl2005a.c:
  * The 'maxdl' variable has fixed value and is used as unsigned in comparisons so change the type to unsigned.
  * Cast 'ret' value to unsigned in comparison with 'size' as previously 'ret' is checked to be not negative and current unsigned comparison behavior is kept.
* library.c: Resolves signed-unsigned compare issues. The variable 'b' is actually unsigned from 'jl2005a_get_pic_data_size' and it is already used as unsigned in compares. Variables 'i' and 'j' are also used as unsigned in 'for' loop comparisons and cannot be negative in array indices.

+++ camlibs/jl2005a/jl2005a.c
@@ -121,7 +121,7 @@ jl2005a_read_picture_data (Camera *camera, GPPort *port,
 					unsigned char *data, unsigned int size)
 {
 	unsigned char *to_read;
-	int maxdl = 0xfa00;
+	unsigned int maxdl = 0xfa00;
 	int ret;
 
 	to_read=data;
@@ -156,7 +156,7 @@ jl2005a_read_picture_data (Camera *camera, GPPort *port,
 	}
 	ret = gp_port_read(port, (char *)to_read, size);
 	if (ret < GP_OK) return ret;
-	if (ret < size) return GP_ERROR;
+	if ((unsigned int)ret < size) return GP_ERROR;
 	/* Switch the inep back to 0x84. */
 	set_usb_in_endpoint	(camera, 0x84);
 	return GP_OK;
